Add lazy provider access
Mostly motivated by the desire to use this library within a
legacy-dapp-support extension, which needs to support the case where
this script could be run before the provider-providing extension (like
MetaMask) is run.

In that case, the web3 returned by this script needs to _lazily_ access
its provider, and so rather than pass it a provider directly, I'm now
constructing it with a Proxy that only accesses the global provider on a
per-access basis.